+++ src/compas_fab/robots/ros_fileserver_loader.py
@@ -5,15 +5,18 @@ from __future__ import print_function
 import binascii
 import logging
 import os
+import tempfile
 
 import roslibpy
 from compas.datastructures import Mesh
+from compas.datastructures import mesh_transform
+from compas.datastructures import meshes_join
 from compas.files import XML
 from compas.geometry import Frame
 from compas.geometry import Transformation
-from compas.datastructures import mesh_transform
-
 from compas.robots.resources.basic import _get_file_format
+from compas.robots.resources.basic import _mesh_import
+from compas.utilities import await_callback
 
 LOGGER = logging.getLogger('compas_fab.robots.ros')
 
@@ -23,177 +26,109 @@ __all__ = [
 
 
 class RosFileServerLoader(object):
-    """Allows to retrieve the mesh files specified in the robot urdf from the
-    ROS file_server, stores it on the local file system and allows to load the
-    meshes afterwards.
+    """Allows to retrieve the mesh files specified in the robot model from the
+    ROS File Server. Optionally, it stores them on the local file system,
+    allowing them to be loaded by local package loaders afterwards.
 
-    It is implemented similar to
-    https://github.com/siemens/ros-sharp/blob/master/Libraries/RosBridgeClient/UrdfImporter.cs
-
-    Attributes:
-        ros (:class:`Ros`): The ROS client
-        local_directory (str, optional): A directory to store the files.
-            Defaults to ~/robot_description
-        robot_name (str or None): The name of the robot, will be read from the
-            urdf
-        requested_resource_files (dict): The mesh files from the urdf
-        status (dict): To check the status of the process.
+    Parameters
+    ----------
+    ros : :class:`compas_fab.backends.RosClient`
+         The ROS client.
+    local_cache : bool
+        ``True`` to store a local copy of the ROS files, otherwise ``False``.
+        Defaults to ``False``.
+    local_cache_directory : str, optional
+        Directory name to store the cached files. Only used if
+        ``local_cache`` is ``True``. Defaults to ``~/robot_description``.
     """
 
-    def __init__(self, ros=None, local_directory=None):
-        self.ros = ros
-        self.local_directory = local_directory or os.path.join(os.path.expanduser('~'), 'robot_description')
-        self.robot_name = None
-        self.requested_resource_files = {}
-        self.status = {"robot_description_received": False,
-                       "robot_description_semantic_received": False,
-                       "resource_files_received": False,
-                       }
-
+    def __init__(self, ros=None, local_cache=False, local_cache_directory=None):
+        self._robot_name = None
         self.schema_prefix = 'package://'
+        self.ros = ros
+        self.local_cache_directory = None
+        self.local_cache_enabled = local_cache
 
-    @classmethod
-    def from_robot_resource_path(cls, path, ros=None):
-        local_directory = os.path.abspath(os.path.join(path, ".."))
-        importer = cls(ros, local_directory)
-        importer.robot_name = os.path.basename(os.path.normpath(path))
-        return importer
-
-    @classmethod
-    def from_urdf_model(cls, urdf_model, ros=None):
-        importer = cls(ros, None)
-        importer.robot_name = os.path.basename(urdf_model.name)
-        return importer
+        if self.local_cache_enabled:
+            self.local_cache_directory = local_cache_directory or os.path.join(
+                os.path.expanduser('~'), 'robot_description')
 
     @property
-    def robot_resource_path(self):
-        if self.robot_name:
-            return os.path.join(self.local_directory, self.robot_name)
-        else:
-            return None # or error
+    def _robot_resource_path(self):
+        if not self._robot_name:
+            raise Exception('Robot name is not assigned, make sure you loaded URDF first')
 
-    def robot_resource_filename(self, resource_file_uri):
-        return os.path.abspath(os.path.join(self.robot_resource_path, resource_file_uri[len('package://'):]))
-
-    def receive_robot_description(self, robot_description):
-        robot_name, uris = self.read_robot_name_and_uris_from_urdf(robot_description)
-        self.robot_name = robot_name
-        # Import resource files
-        self.import_resource_files(uris)
-        # Save robot_description.urdf
-        self.save_robot_description(robot_description)
-        # Save robot_description_semantic.urdf
-        param = roslibpy.Param(self.ros, '/robot_description_semantic')
-        param.get(self.save_robot_description_semantic)
+        return os.path.join(self.local_cache_directory, self._robot_name)
 
     @property
-    def urdf_filename(self):
-        return os.path.join(self.robot_resource_path, "urdf", "robot_description.urdf")
+    def _urdf_filename(self):
+        return os.path.join(self._robot_resource_path, 'urdf', 'robot_description.urdf')
 
     @property
-    def srdf_filename(self):
-        return os.path.join(self.robot_resource_path, "robot_description_semantic.srdf")
+    def _srdf_filename(self):
+        return os.path.join(self._robot_resource_path, 'robot_description_semantic.srdf')
 
-    def save_robot_description(self, robot_description):
-        # Save robot_description.urdf
-        filename = self.urdf_filename
-        LOGGER.info("Saving URDF file to %s" % filename)
-        self.write_file(filename, robot_description)
-        self.status.update({"robot_description_received": True})
+    def load_urdf(self, parameter_name='/robot_description'):
+        """Loads a URDF model from the specified ROS parameter.
 
-    def save_robot_description_semantic(self, robot_description_semantic):
-        # Save robot_description_semantic.urdf
-        filename = self.srdf_filename
-        LOGGER.info("Saving URDF file to %s" % filename)
-        self.write_file(filename, robot_description_semantic)
-        self.status.update({"robot_description_semantic_received": True})
+        Parameters
+        ----------
+        parameter_name : str, optional
+            Name of the ROS parameter containing the robot description.
+            Defaults to ``/robot_description``.
 
-    def load(self):
-        param = roslibpy.Param(self.ros, '/robot_description')
-        param.get(self.receive_robot_description)
+        Returns
+        -------
+        str
+            URDF model of the robot currently loaded in ROS.
+        """
+        param = roslibpy.Param(self.ros, parameter_name)
+        urdf = await_callback(param.get)
+        self._robot_name = self._read_robot_name(urdf)
 
-    def receive_resource_file(self, local_filename, resource_file_uri):
+        if self.local_cache_enabled:
+            self._write_file(self._urdf_filename, urdf)
 
-        def write_binary_response_to_file(response):
-            LOGGER.info("Saving %s to %s" % (resource_file_uri, local_filename))
-            self.write_file(local_filename, binascii.a2b_base64(response.data['value']), 'wb')
-            #self.write_file(local_filename, response.data['value'])
-            self.update_file_request_status(resource_file_uri)
+        return urdf
 
-        service = roslibpy.Service(self.ros, "/file_server/get_file",
-                                   "file_server/GetBinaryFile")
-        service.call(roslibpy.ServiceRequest({'name': resource_file_uri}),
-                     write_binary_response_to_file, errback=None)
+    def load_srdf(self, parameter_name='/robot_description_semantic'):
+        """Loads an SRDF model from the specified ROS parameter.
 
-    def read_robot_name_and_uris_from_urdf(self, robot_description):
+        Parameters
+        ----------
+        parameter_name : str, optional
+            Name of the ROS parameter containing the robot semantics.
+            Defaults to ``/robot_description_semantic``.
+
+        Returns
+        -------
+        str
+            SRDF model of the robot currently loaded in ROS.
+        """
+        param = roslibpy.Param(self.ros, parameter_name)
+        srdf = await_callback(param.get)
+
+        if self.local_cache_enabled:
+            self._write_file(self._srdf_filename, srdf)
+
+        return srdf
+
+    def _read_robot_name(self, robot_description):
+        # TODO: Optimize this. We really don't need to parse the full URDF
+        # only to read the robot's name (only used for local caching)
         xml = XML.from_string(robot_description)
-        robot_name = xml.root.attrib['name']
-        uris = [mesh.attrib['filename'] for mesh in xml.root.iter('mesh')
-            if mesh.attrib['filename'] != '']
-        return robot_name, uris
+        return xml.root.attrib['name']
 
-    def import_resource_files(self, uris):
-        for resource_file_uri in uris:
-            self.requested_resource_files.update({resource_file_uri: False})
-            local_filename = self.robot_resource_filename(resource_file_uri)
-            self.receive_resource_file(local_filename, str(resource_file_uri))
+    def _write_file(self, filename, file_contents, mode='w'):
+        LOGGER.debug('Saving file to %s', filename)
 
-    def write_file(self, filename, filecontents, mode='w'):
         dirname = os.path.dirname(filename)
+
         if not os.path.isdir(dirname):
             os.makedirs(dirname)
-        f = open(filename, mode)
-        f.write(filecontents)
-        f.close()
 
-    def update_file_request_status(self, resource_file_uri):
-        self.requested_resource_files[resource_file_uri] = True
-        if all(self.requested_resource_files.values()):
-            self.status.update({"resource_files_received": True})
-
-    def read_mesh_from_resource_file_uri(self, resource_file_uri, meshcls):
-        """Reads the mesh from a file uri and creates a mesh type based on the
-        passed mesh class.
-
-        Args:
-            resource_file_uri (str): The resourec file starting with package://
-            meshcls (:class:): A class that allows to create a custom mesh type
-                and is created with a :class:`Mesh`
-        """
-        filename = self.robot_resource_filename(resource_file_uri)
-        return self.read_mesh_from_filename(filename, meshcls)
-
-    def read_mesh_from_filename(self, filename, meshcls):
-        if not os.path.isfile(filename):
-            raise FileNotFoundError("No such file: '%s'" % filename)
-        extension = filename[(filename.rfind(".") + 1):]
-        if extension == "dae": # no dae support yet
-            #mesh = Mesh.from_dae(filename)
-            obj_filename = filename.replace(".dae", ".obj")
-            if os.path.isfile(obj_filename):
-                mesh = Mesh.from_obj(obj_filename)
-                # former DAE files have yaxis and zaxis swapped
-                # TODO: already fix in conversion to obj
-                frame = Frame([0,0,0], [1,0,0], [0,0,1])
-                T = Transformation.from_frame(frame)
-                mesh_transform(mesh, T)
-            else:
-                raise FileNotFoundError("Please convert '%s' into an OBJ file, \
-                                         since DAE is currently not supported \
-                                         yet." % filename)
-        elif extension == "obj":
-            mesh = Mesh.from_obj(filename)
-        elif extension == "stl":
-            mesh = Mesh.from_stl(filename)
-        else:
-            raise ValueError("%s file types not yet supported" %
-                extension.upper())
-
-        return meshcls(mesh)
-
-    def _get_local_path(self, url):
-        _prefix, path = url.split(self.schema_prefix)
-        return os.path.abspath(os.path.join(self.robot_resource_path, path))
+        with open(filename, mode) as f:
+            f.write(file_contents)
 
     def can_load_mesh(self, url):
         """Determine whether this loader can load a given mesh URL.
@@ -206,15 +141,10 @@ class RosFileServerLoader(object):
         Returns
         -------
         bool
-            ``True`` if the URL uses the ``package://` scheme and the package name
-            matches the specified in the constructor and the file exists locally,
+            ``True`` if the URL uses the ``package://` scheme,
             otherwise ``False``.
         """
-        if not url.startswith(self.schema_prefix):
-            return False
-
-        local_file = self._get_local_path(url)
-        return os.path.isfile(local_file)
+        return url.startswith(self.schema_prefix)
 
     def load_mesh(self, url):
         """Loads a mesh from local storage.
@@ -229,49 +159,81 @@ class RosFileServerLoader(object):
         :class:`Mesh`
             Instance of a mesh.
         """
-        local_file = self._get_local_path(url)
-        return _mesh_import(url, local_file)
+
+        service = roslibpy.Service(self.ros, '/file_server/get_file', 'file_server/GetBinaryFile')
+        request = roslibpy.ServiceRequest(dict(name=url))
+        response = await_callback(service.call, 'callback', 'errback', request)
+
+        file_extension = _get_file_format(url)
+        file_content = binascii.a2b_base64(response.data['value'])
+        if self.local_cache_enabled:
+            local_filename = self._local_mesh_filename(url)
+        else:
+            _, local_filename = tempfile.mkstemp(suffix='.' + file_extension, prefix='ros_fileserver_')
+
+        # Just look away, we're about to do something nasty!
+        # namespaces are handled differently between the CLI and CPython
+        # XML parsers, so, we just get rid of it for DAE files
+        if file_extension == 'dae':
+            file_content = file_content.replace(b'xmlns="http://www.collada.org/2005/11/COLLADASchema"', b'')
+
+        # compas.files does not support file-like objects so we need to
+        # save the file to disk always. If local caching is enabled,
+        # we store it in the cache folder, otherwise, as a temp file.
+        self._write_file(local_filename, file_content, 'wb')
+
+        return _fileserver_mesh_import(url, local_filename)
+
+    def _local_mesh_filename(self, url):
+        return os.path.abspath(os.path.join(self._robot_resource_path, url[len('package://'):]))
 
 
-SUPPORTED_FORMATS = ('obj', 'stl', 'ply', 'dae')
+def _dae_mesh_importer(filename):
+    """This is a very simple implementation of a DAE/Collada parser.
+    It merges all solids of the DAE file into one mesh, because
+    several other parts of the framework don't support multi-meshes per file."""
+    dae = XML.from_file(filename)
+    meshes = []
+
+    for mesh_xml in dae.root.findall('.//mesh'):
+        for triangle_set in mesh_xml.findall('triangles'):
+            triangle_set_data = triangle_set.find('p').text.split()
+
+            # Parse vertices
+            vertices_input = triangle_set.find('input[@semantic="VERTEX"]')
+            vertices_link = mesh_xml.find('vertices[@id="{}"]/input'.format(vertices_input.attrib['source'][1:]))
+            positions = mesh_xml.find('source[@id="{}"]/float_array'.format(vertices_link.attrib['source'][1:]))
+            positions = positions.text.split()
+
+            vertices = list(map(float, positions[i:i + 3]) for i in range(0, len(positions), 3))
+
+            # Parse faces
+            faces = list(map(int, triangle_set_data[::2]))  # Ignore normals (ever second item is normal index)
+            faces = list(faces[i:i + 3] for i in range(0, len(faces), 3))
+
+            mesh = Mesh.from_vertices_and_faces(vertices, faces)
+
+            meshes.append(mesh)
+
+    combined_mesh = meshes_join(meshes)
+    # former DAE files have yaxis and zaxis swapped
+    # frame = Frame([0, 0, 0], [1, 0, 0], [0, 0, 1])
+    # T = Transformation.from_frame(frame)
+    # mesh_transform(combined_mesh, T)
+    return combined_mesh
 
 
-def _mesh_import(url, filename):
-    """Internal function to load meshes using the correct loader.
-
-    Name and file might be the same but not always, e.g. temp files."""
+def _fileserver_mesh_import(url, filename):
+    """Internal function that adds primitive support for DAE files
+    to the _mesh_import function of compas.robots."""
     file_extension = _get_file_format(url)
 
-    if file_extension not in SUPPORTED_FORMATS:
-        raise NotImplementedError(
-            'Mesh type not supported: {}'.format(file_extension))
+    if file_extension == 'dae':
+        # Magic!
+        return _dae_mesh_importer(filename)
+    else:
+        return _mesh_import(url, filename)
 
-    print(filename)
-
-    if file_extension == "dae": # no dae support yet
-        #mesh = Mesh.from_dae(filename)
-        obj_filename = filename.replace(".dae", ".obj")
-        if os.path.isfile(obj_filename):
-            mesh = Mesh.from_obj(obj_filename)
-            # former DAE files have yaxis and zaxis swapped
-            # TODO: already fix in conversion to obj
-            frame = Frame([0,0,0], [1,0,0], [0,0,1])
-            T = Transformation.from_frame(frame)
-            mesh_transform(mesh, T)
-            return mesh
-        else:
-            raise FileNotFoundError("Please convert '%s' into an OBJ file, \
-                                        since DAE is currently not supported \
-                                        yet." % filename)
-
-    if file_extension == 'obj':
-        return Mesh.from_obj(filename)
-    elif file_extension == 'stl':
-        return Mesh.from_stl(filename)
-    elif file_extension == 'ply':
-        return Mesh.from_ply(filename)
-
-    raise Exception
 
 if __name__ == "__main__":
 
@@ -296,4 +258,3 @@ if __name__ == "__main__":
     ros.call_later(50, ros.close)
     ros.call_later(52, ros.terminate)
     ros.run_forever()
-